mustafa/requirements.txt
# PyTorch CPU Version
torch>=2.0.0
torchvision>=0.15.0

# COCO Evaluation
pycocotools>=2.0.6

# Image Processing
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# Data Science
numpy>=1.21.0
